extract search and list into own components
