+++ packages/components/src/Combobox/components/ComboboxContent/ComboboxContent.css.d.ts
@@ -1,13 +1,7 @@
 declare const styles: {
   readonly "content": string;
   readonly "hidden": string;
-  readonly "optionsList": string;
-  readonly "option": string;
-  readonly "search": string;
-  readonly "searchInput": string;
-  readonly "clearSearch": string;
   readonly "actions": string;
-  readonly "selectedOption": string;
   readonly "actionPadding": string;
 };
 export = styles;
