+++ packages/components/src/Combobox/components/ComboboxContent/ComboboxList.css.d.ts
@@ -1,0 +1,7 @@
+declare const styles: {
+  readonly "optionsList": string;
+  readonly "option": string;
+  readonly "selectedOption": string;
+};
+export = styles;
+
