+++ packages/components/src/Combobox/components/ComboboxContent/ComboboxSearch.css.d.ts
@@ -1,0 +1,7 @@
+declare const styles: {
+  readonly "search": string;
+  readonly "searchInput": string;
+  readonly "clearSearch": string;
+};
+export = styles;
+
